[RNKC-080] - add dynamic comparison example
