+++ example/src/screens/Examples/ReanimatedChat/hooks.ts
@@ -1,0 +1,39 @@
+import { Platform } from 'react-native';
+import { useKeyboardHandler } from 'react-native-keyboard-controller';
+import { Easing, useSharedValue, withTiming } from 'react-native-reanimated';
+
+export const useTelegramTransitions = () => {
+  const height = useSharedValue(0);
+
+  useKeyboardHandler(
+    {
+      onStart: (e) => {
+        'worklet';
+
+        if (Platform.OS === 'android') {
+          // on Android Telegram is not using androidx.core values and uses custom interpolation
+          // duration is taken from here: https://github.com/DrKLO/Telegram/blob/e9a35cea54c06277c69d41b8e25d94b5d7ede065/TMessagesProj/src/main/java/org/telegram/ui/ActionBar/AdjustPanLayoutHelper.java#L39
+          // and bezier is taken from: https://github.com/DrKLO/Telegram/blob/e9a35cea54c06277c69d41b8e25d94b5d7ede065/TMessagesProj/src/main/java/androidx/recyclerview/widget/ChatListItemAnimator.java#L40
+          height.value = withTiming(-e.height, {
+            duration: 250,
+            easing: Easing.bezier(
+              0.19919472913616398,
+              0.010644531250000006,
+              0.27920937042459737,
+              0.91025390625
+            ),
+          });
+        } else {
+          // on iOS Telegram simply moves TextInput synchronously with the content
+          // to achieve such behavior we are instantly change `height.value` to keyboard
+          // final frame - iOS will schedule layout animation and it will move the content
+          // altogether with the keyboard
+          height.value = -e.height;
+        }
+      },
+    },
+    []
+  );
+
+  return { height };
+};
